requirements.txt
numpy
pytorch
gym
opencv-python
matplotlib
tensorflow
keras